Changed name of NotesController's file (Controller.cs) to NotesController.cs
